requirements.txt
future~=0.18.2
setuptools~=58.1.0
fitz~=0.0.1.dev2
Pillow~=8.3.1
Pygments~=2.10.0
cefpython3~=66.1